+++ pkg/project/resource/resource_project_environment_test.go
@@ -65,7 +65,7 @@ func TestAccProjectEnvironment_UpgradeFromSDKv2(t *testing.T) {
 			{
 				ProtoV6ProviderFactories: acctest.ProtoV6ProviderFactories,
 				Config:                   config,
-				PlanOnly:                 true,
+				PlanOnly:                 false,
 				ConfigPlanChecks:         testutil.ConfigPlanChecks(fqrn),
 			},
 		},