+++ pkg/project/resource/resource_project_role_test.go
@@ -75,7 +75,7 @@ func TestAccProjectRole_UpgradeFromSDKv2(t *testing.T) {
 			{
 				ProtoV6ProviderFactories: acctest.ProtoV6ProviderFactories,
 				Config:                   config,
-				PlanOnly:                 true,
+				PlanOnly:                 false,
 				ConfigPlanChecks:         testutil.ConfigPlanChecks(fqrn),
 			},
 		},